feat: introduce @arkenv/build-utils shared package and refactor nextjs and nuxt integrations
