+++ packages/build-utils/package.json
@@ -1,0 +1,39 @@
+{
+	"name": "@arkenv/build-utils",
+	"version": "1.0.0-alpha.0",
+	"description": "Shared build and codegen utilities for ArkEnv framework plugins",
+	"type": "module",
+	"main": "./dist/index.cjs",
+	"module": "./dist/index.js",
+	"types": "./dist/index.d.ts",
+	"exports": {
+		"types": "./dist/index.d.ts",
+		"import": "./dist/index.js",
+		"require": "./dist/index.cjs"
+	},
+	"scripts": {
+		"build": "tsdown",
+		"typecheck": "tsc --noEmit",
+		"clean": "rimraf dist node_modules"
+	},
+	"files": [
+		"dist"
+	],
+	"dependencies": {
+		"chokidar": "^4.0.3"
+	},
+	"devDependencies": {
+		"@types/node": "catalog:",
+		"tsdown": "catalog:",
+		"typescript": "catalog:"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/yamcodes/arkenv.git"
+	},
+	"bugs": "https://github.com/yamcodes/arkenv/issues"
+}
