+++ packages/nextjs/package.json
@@ -14,6 +14,7 @@
 		"server-only": "^0.0.1"
 	},
 	"devDependencies": {
+		"@arkenv/build-utils": "workspace:*",
 		"@repo/scope": "workspace:*",
 		"@repo/types": "workspace:*",
 		"@size-limit/preset-small-lib": "catalog:",
