+++ packages/nextjs/src/config.ts
@@ -1,11 +1,20 @@
 import fs from "node:fs";
 import path from "node:path";
-import { watch as chokidarWatch } from "chokidar";
+import {
+	findSchemaPath,
+	resolveLayout,
+	watchSchema,
+	extractKeys,
+	extractClientKeys,
+	extractSharedKeys,
+} from "@arkenv/build-utils";
 
-declare global {
-	// eslint-disable-next-line no-var
-	var __arkenv_watcher__: import("chokidar").FSWatcher | undefined;
-}
+// Re-export key extraction utilities for backwards compatibility and testing
+export {
+	extractKeys,
+	extractClientKeys,
+	extractSharedKeys,
+} from "@arkenv/build-utils";
 
 /**
  * Configuration options for the ArkEnv Next.js integration.
@@ -136,190 +145,15 @@ export function withArkEnv<T>(nextConfig: T, options?: ArkEnvConfigOptions): T {
 						path.join(baseDir, "server.ts"),
 					].filter(fs.existsSync)
 				: [schemaPath];
-		watchSchema(watchPaths, outputPath, resolvedLayout);
+		watchSchema(watchPaths, () => {
+			const mainSchemaPath = Array.isArray(watchPaths)
+				? watchPaths[0]
+				: watchPaths;
+			runCodegen(mainSchemaPath, outputPath, resolvedLayout);
+		});
 	}
 
 	return nextConfig;
-}
-
-/**
- * Find the path to the schema file in the project.
- *
- * @returns The absolute path to the schema file, or null if not found
- */
-/**
- * Resolve the layout (simple vs strict) and the base directory from a schema path.
- *
- * Auto-detects layout when `layoutOption` is not provided. When `layoutOption`
- * is `"strict"`, validates that the required split files exist and throws a
- * descriptive error if either is missing.
- *
- * @param schemaPath The absolute path to the schema file or directory
- * @param layoutOption The explicit layout option, if provided
- * @returns The resolved layout and base directory
- * @throws An error if strict layout files are missing
- */
-function resolveLayout(
-	schemaPath: string,
-	layoutOption?: "simple" | "strict",
-): { layout: "simple" | "strict"; baseDir: string } {
-	const checkStrict = (dir: string) =>
-		fs.existsSync(path.join(dir, "internal", "shared.ts")) &&
-		fs.existsSync(path.join(dir, "client.ts")) &&
-		fs.existsSync(path.join(dir, "server.ts"));
-
-	const resolveBaseDir = (p: string): string => {
-		// Normalize: if schemaPath has an extension, try the extensionless form as a dir
-		const ext = path.extname(p);
-		const baseWithoutExt = ext ? p.slice(0, -ext.length) : p;
-		if (
-			fs.existsSync(baseWithoutExt) &&
-			fs.statSync(baseWithoutExt).isDirectory()
-		) {
-			return baseWithoutExt;
-		}
-		return p;
-	};
-
-	if (!layoutOption) {
-		// Auto-detect
-		const resolved = resolveBaseDir(schemaPath);
-		if (fs.existsSync(resolved) && fs.statSync(resolved).isDirectory()) {
-			if (checkStrict(resolved)) {
-				return { layout: "strict", baseDir: resolved };
-			}
-			return { layout: "simple", baseDir: resolved };
-		}
-
-		// schemaPath is a file — check surrounding dirs
-		const parent = path.dirname(schemaPath);
-		const ext = path.extname(schemaPath);
-		const baseWithoutExt = ext ? schemaPath.slice(0, -ext.length) : schemaPath;
-		if (
-			fs.existsSync(baseWithoutExt) &&
-			fs.statSync(baseWithoutExt).isDirectory() &&
-			checkStrict(baseWithoutExt)
-		) {
-			return { layout: "strict", baseDir: baseWithoutExt };
-		}
-		if (checkStrict(parent)) {
-			return { layout: "strict", baseDir: parent };
-		}
-		if (
-			path.basename(parent) === "internal" &&
-			checkStrict(path.dirname(parent))
-		) {
-			return { layout: "strict", baseDir: path.dirname(parent) };
-		}
-		return { layout: "simple", baseDir: schemaPath };
-	}
-
-	if (layoutOption === "strict") {
-		// Resolve baseDir for an explicit strict layout
-		let baseDir: string;
-		const resolved = resolveBaseDir(schemaPath);
-		if (fs.existsSync(resolved) && fs.statSync(resolved).isDirectory()) {
-			baseDir = resolved;
-		} else {
-			const parent = path.dirname(schemaPath);
-			if (path.basename(parent) === "internal") {
-				baseDir = path.dirname(parent);
-			} else {
-				baseDir = parent;
-			}
-		}
-
-		// Validate that required split files exist
-		const clientPath = path.join(baseDir, "client.ts");
-		const sharedPath = path.join(baseDir, "internal", "shared.ts");
-		if (!fs.existsSync(clientPath) || !fs.existsSync(sharedPath)) {
-			throw new Error(
-				`[ArkEnv] Strict layout requires "${clientPath}" and "${sharedPath}" to exist. ` +
-					`Ensure both files are present or remove the 'layout: "strict"' option to let ArkEnv auto-detect.`,
-			);
-		}
-
-		return { layout: "strict", baseDir };
-	}
-
-	// layoutOption === "simple"
-	return { layout: "simple", baseDir: schemaPath };
-}
-
-function findSchemaPath(): string | null {
-	const possiblePaths = [
-		path.join(process.cwd(), "src", "env.ts"),
-		path.join(process.cwd(), "env.ts"),
-	];
-	for (const p of possiblePaths) {
-		if (fs.existsSync(p)) return p;
-	}
-
-	const possibleDirs = [
-		path.join(process.cwd(), "src", "env"),
-		path.join(process.cwd(), "env"),
-	];
-	for (const d of possibleDirs) {
-		if (
-			fs.existsSync(d) &&
-			fs.existsSync(path.join(d, "internal", "shared.ts")) &&
-			fs.existsSync(path.join(d, "client.ts")) &&
-			fs.existsSync(path.join(d, "server.ts"))
-		) {
-			return d;
-		}
-	}
-	return null;
-}
-
-/**
- * Watch the schema file for changes and trigger codegen.
- *
- * @param schemaPath The absolute path to the schema file, or an array of paths to watch
- * @param outputPath The absolute path to the generated output file
- * @param layout The resolved layout to pass through to codegen on each change
- */
-function watchSchema(
-	schemaPath: string | string[],
-	outputPath: string,
-	layout?: "simple" | "strict",
-) {
-	const previousWatcher = globalThis.__arkenv_watcher__;
-	if (previousWatcher && typeof previousWatcher.close === "function") {
-		previousWatcher.close().catch((err: unknown) => {
-			const message = err instanceof Error ? err.message : String(err);
-			// biome-ignore lint/suspicious/noConsole: watcher errors must be logged
-			console.error(
-				`[ArkEnv Watcher] Failed to close previous watcher: ${message}`,
-			);
-		});
-	}
-
-	try {
-		const watcher = chokidarWatch(schemaPath, { ignoreInitial: true });
-		globalThis.__arkenv_watcher__ = watcher;
-
-		watcher.on("change", () => {
-			try {
-				const mainSchemaPath = Array.isArray(schemaPath)
-					? schemaPath[0]
-					: schemaPath;
-				runCodegen(mainSchemaPath, outputPath, layout);
-			} catch (err: unknown) {
-				const message = err instanceof Error ? err.message : String(err);
-				// biome-ignore lint/suspicious/noConsole: watcher errors must be logged
-				console.error(
-					`[ArkEnv Watcher] Failed to regenerate env.gen.ts: ${message}`,
-				);
-			}
-		});
-	} catch (err: unknown) {
-		const message = err instanceof Error ? err.message : String(err);
-		// biome-ignore lint/suspicious/noConsole: watcher errors must be logged
-		console.error(
-			`[ArkEnv Watcher] Failed to start watch on ${schemaPath}: ${message}`,
-		);
-	}
 }
 
 /**
@@ -380,208 +214,6 @@ export function runCodegen(
 	if (shouldWrite) {
 		fs.writeFileSync(outputPath, generatedCode, "utf-8");
 	}
-}
-
-/**
- * Statically extract client and shared keys from the schema content.
- *
- * @param content The schema file string content
- * @returns An object containing the extracted client and shared keys
- */
-export function extractKeys(content: string): {
-	clientKeys: string[];
-	sharedKeys: string[];
-} {
-	const clientKeys: string[] = [];
-	const sharedKeys: string[] = [];
-
-	// Extract client block
-	const clientBlock = extractBlock(content, "client");
-	if (clientBlock) {
-		clientKeys.push(...parseBlockKeys(clientBlock));
-	}
-
-	// Extract shared block
-	const sharedBlock = extractBlock(content, "shared");
-	if (sharedBlock) {
-		sharedKeys.push(...parseBlockKeys(sharedBlock));
-	}
-
-	return { clientKeys, sharedKeys };
-}
-
-/**
- * Extract a specific block from the schema content by name.
- *
- * @param content The schema file string content
- * @param blockName The name of the block to extract (e.g., 'client' or 'shared')
- * @returns The contents of the block, or null if not found
- */
-function extractBlock(content: string, blockName: string): string | null {
-	// Find "\bblockName\s*:\s*{" or "\bblockName\s*:\s*type({"
-	const regex = new RegExp(
-		`\\b${blockName}\\s*:\\s*(?:[a-zA-Z0-9_$.]+\\s*\\(\\s*)?\\{`,
-		"g",
-	);
-	const match = regex.exec(content);
-	if (!match) return null;
-
-	const startIndex = regex.lastIndex;
-	let braceCount = 1;
-	let index = startIndex;
-	let inString: string | null = null;
-	let inComment: "single" | "multi" | null = null;
-
-	while (index < content.length && braceCount > 0) {
-		const char = content[index];
-		const nextChar = content[index + 1];
-
-		if (inComment === "single") {
-			if (char === "\n" || char === "\r") inComment = null;
-			index++;
-			continue;
-		}
-		if (inComment === "multi") {
-			if (char === "*" && nextChar === "/") {
-				inComment = null;
-				index += 2;
-				continue;
-			}
-			index++;
-			continue;
-		}
-
-		if (inString) {
-			if (char === inString && content[index - 1] !== "\\") {
-				inString = null;
-			}
-			index++;
-			continue;
-		}
-
-		if (char === "/" && nextChar === "/") {
-			inComment = "single";
-			index += 2;
-			continue;
-		}
-		if (char === "/" && nextChar === "*") {
-			inComment = "multi";
-			index += 2;
-			continue;
-		}
-		if (char === "'" || char === '"' || char === "`") {
-			inString = char;
-			index++;
-			continue;
-		}
-
-		if (char === "{") {
-			braceCount++;
-		} else if (char === "}") {
-			braceCount--;
-		}
-		index++;
-	}
-
-	if (braceCount === 0) {
-		return content.slice(startIndex, index - 1);
-	}
-
-	return null;
-}
-
-/**
- * Parse environment variable keys from a block's content.
- *
- * @param blockContent The raw content of the block
- * @returns An array of parsed environment variable keys
- */
-function parseBlockKeys(blockContent: string): string[] {
-	const keys: string[] = [];
-	let inString: string | null = null;
-	let inComment: "single" | "multi" | null = null;
-	let currentToken = "";
-	let lastStringContent = "";
-	let braceDepth = 0;
-
-	for (let i = 0; i < blockContent.length; i++) {
-		const char = blockContent[i];
-		const nextChar = blockContent[i + 1];
-
-		if (inComment === "single") {
-			if (char === "\n" || char === "\r") inComment = null;
-			continue;
-		}
-		if (inComment === "multi") {
-			if (char === "*" && nextChar === "/") {
-				inComment = null;
-				i++;
-			}
-			continue;
-		}
-
-		if (inString) {
-			if (char === inString && blockContent[i - 1] !== "\\") {
-				inString = null;
-				lastStringContent = currentToken;
-				currentToken = "";
-			} else {
-				currentToken += char;
-			}
-			continue;
-		}
-
-		// Start comments/strings
-		if (char === "/" && nextChar === "/") {
-			inComment = "single";
-			i++;
-			continue;
-		}
-		if (char === "/" && nextChar === "*") {
-			inComment = "multi";
-			i++;
-			continue;
-		}
-		if (char === "'" || char === '"' || char === "`") {
-			inString = char;
-			currentToken = "";
-			continue;
-		}
-
-		if (char === "{") {
-			braceDepth++;
-			currentToken = "";
-			lastStringContent = "";
-			continue;
-		}
-		if (char === "}") {
-			braceDepth--;
-			currentToken = "";
-			lastStringContent = "";
-			continue;
-		}
-
-		if (char === ":") {
-			if (braceDepth === 0) {
-				const key = currentToken.trim() || lastStringContent.trim();
-				if (key) {
-					keys.push(key);
-				}
-			}
-			currentToken = "";
-			lastStringContent = "";
-			continue;
-		}
-
-		if (/[a-zA-Z0-9_$]/.test(char)) {
-			currentToken += char;
-		} else if (char === "," || char === "\n" || char === "\r") {
-			currentToken = "";
-			lastStringContent = "";
-		}
-	}
-
-	return keys;
 }
 
 /**
@@ -692,190 +324,4 @@ ${runtimeEnvLines}
 const arkenv = createEnv;
 export default arkenv;
 `;
-}
-
-/**
- * Extract env var keys from a strict-layout `client.ts` file.
- *
- * Locates the first `arkenv({...})` call and delegates to `parseBlockKeys`
- * to enumerate the keys defined in the schema object.
- *
- * @param content The source text of `client.ts`
- * @returns An array of env var key names found in the client schema
- */
-export function extractClientKeys(content: string): string[] {
-	const block = extractClientBlock(content);
-	return block ? parseBlockKeys(block) : [];
-}
-
-/**
- * Extract env var keys from a strict-layout `internal/shared.ts` file.
- *
- * Locates the `SharedSchema = ...({...})` assignment and delegates to
- * `parseBlockKeys` to enumerate the keys defined in the schema object.
- *
- * @param content The source text of `internal/shared.ts`
- * @returns An array of env var key names found in the shared schema
- */
-export function extractSharedKeys(content: string): string[] {
-	const block = extractSharedBlock(content);
-	return block ? parseBlockKeys(block) : [];
-}
-
-/**
- * Extract the schema object body from a strict-layout `client.ts` file.
- *
- * Matches the first `arkenv(...)` call, then performs brace-balanced
- * traversal while respecting strings and comments to find the boundaries
- * of the first argument object.
- *
- * @param content The source text of `client.ts`
- * @returns The raw text inside the first `{...}` argument, or `null` if not found
- */
-function extractClientBlock(content: string): string | null {
-	const regex = /\barkenv\s*\(\s*(?:[a-zA-Z0-9_$.]+\s*\(\s*)*\{/g;
-	const match = regex.exec(content);
-	if (!match) return null;
-
-	const startIndex = regex.lastIndex;
-	let braceCount = 1;
-	let index = startIndex;
-	let inString: string | null = null;
-	let inComment: "single" | "multi" | null = null;
-
-	while (index < content.length && braceCount > 0) {
-		const char = content[index];
-		const nextChar = content[index + 1];
-
-		if (inComment === "single") {
-			if (char === "\n" || char === "\r") inComment = null;
-			index++;
-			continue;
-		}
-		if (inComment === "multi") {
-			if (char === "*" && nextChar === "/") {
-				inComment = null;
-				index += 2;
-				continue;
-			}
-			index++;
-			continue;
-		}
-
-		if (inString) {
-			if (char === inString && content[index - 1] !== "\\") {
-				inString = null;
-			}
-			index++;
-			continue;
-		}
-
-		if (char === "/" && nextChar === "/") {
-			inComment = "single";
-			index += 2;
-			continue;
-		}
-		if (char === "/" && nextChar === "*") {
-			inComment = "multi";
-			index += 2;
-			continue;
-		}
-		if (char === "'" || char === '"' || char === "`") {
-			inString = char;
-			index++;
-			continue;
-		}
-
-		if (char === "{") {
-			braceCount++;
-		} else if (char === "}") {
-			braceCount--;
-		}
-		index++;
-	}
-
-	if (braceCount === 0) {
-		return content.slice(startIndex, index - 1);
-	}
-
-	return null;
-}
-
-/**
- * Extract the schema object body from a strict-layout `internal/shared.ts` file.
- *
- * Matches the `SharedSchema = ...({...})` assignment, then performs
- * brace-balanced traversal while respecting strings and comments to find
- * the boundaries of the schema object.
- *
- * @param content The source text of `internal/shared.ts`
- * @returns The raw text inside the `{...}` schema object, or `null` if not found
- */
-function extractSharedBlock(content: string): string | null {
-	const regex = /\bSharedSchema\s*=\s*(?:[a-zA-Z0-9_$.]+\s*\(\s*)*\{/g;
-	const match = regex.exec(content);
-	if (!match) return null;
-
-	const startIndex = regex.lastIndex;
-	let braceCount = 1;
-	let index = startIndex;
-	let inString: string | null = null;
-	let inComment: "single" | "multi" | null = null;
-
-	while (index < content.length && braceCount > 0) {
-		const char = content[index];
-		const nextChar = content[index + 1];
-
-		if (inComment === "single") {
-			if (char === "\n" || char === "\r") inComment = null;
-			index++;
-			continue;
-		}
-		if (inComment === "multi") {
-			if (char === "*" && nextChar === "/") {
-				inComment = null;
-				index += 2;
-				continue;
-			}
-			index++;
-			continue;
-		}
-
-		if (inString) {
-			if (char === inString && content[index - 1] !== "\\") {
-				inString = null;
-			}
-			index++;
-			continue;
-		}
-
-		if (char === "/" && nextChar === "/") {
-			inComment = "single";
-			index += 2;
-			continue;
-		}
-		if (char === "/" && nextChar === "*") {
-			inComment = "multi";
-			index += 2;
-			continue;
-		}
-		if (char === "'" || char === '"' || char === "`") {
-			inString = char;
-			index++;
-			continue;
-		}
-
-		if (char === "{") {
-			braceCount++;
-		} else if (char === "}") {
-			braceCount--;
-		}
-		index++;
-	}
-
-	if (braceCount === 0) {
-		return content.slice(startIndex, index - 1);
-	}
-
-	return null;
 }
