+++ packages/nuxt/package.json
@@ -13,6 +13,7 @@
 		"chokidar": "^4.0.3"
 	},
 	"devDependencies": {
+		"@arkenv/build-utils": "workspace:*",
 		"@nuxt/kit": "^3.12.0",
 		"@nuxt/schema": "^3.12.0",
 		"@repo/scope": "workspace:*",
