+++ packages/nuxt/src/config.ts
@@ -5,7 +5,9 @@ export {
 	extractKeys,
 	extractServerKeys,
 	extractSharedKeys,
-} from "./parser";
-export { findSchemaPath, resolveLayout } from "./resolver";
-export type { LayoutMode, Logger, ResolvedLayout } from "./types";
-export { closeWatcher, watchSchema } from "./watcher";
+	findSchemaPath,
+	resolveLayout,
+	closeWatcher,
+	watchSchema,
+} from "@arkenv/build-utils";
+export type { LayoutMode, Logger, ResolvedLayout } from "@arkenv/build-utils";
