+++ packages/nuxt/src/types.ts
@@ -1,4 +1,5 @@
 import type { InferType } from "@repo/types";
+export type { LayoutMode, Logger, ResolvedLayout } from "@arkenv/build-utils";
 
 /**
  * Shared type utilities for Nuxt entry points.
@@ -20,15 +21,3 @@ export type MergeExtends<TExtends extends readonly unknown[] | undefined> =
 	TExtends extends readonly unknown[]
 		? UnionToIntersection<ResolveExtendsElement<TExtends[number]>>
 		: {};
-
-export type LayoutMode = "simple" | "strict";
-
-export type ResolvedLayout = {
-	layout: LayoutMode;
-	baseDir: string;
-};
-
-export type Logger = {
-	error: (msg: string) => void;
-	info?: (msg: string) => void;
-};
